Adiciona o seed da base de dados , endereco e clientes

+++ src/main/java/com/igomarcelino/gerenciamento_ordem_de_servico/entities/Cliente.java
@@ -14,7 +14,7 @@ public class Cliente extends Pessoa{
     public Cliente() {
     }
 
-    public Cliente(Long id, String nome, String cpf, String telefone, String email, Endereco endereco) {
+    public Cliente(Integer id, String nome, String cpf, String telefone, String email, Endereco endereco) {
         super(id, nome, cpf, telefone, email);
         this.endereco = endereco;
     }
