+++ src/main/java/com/igomarcelino/gerenciamento_ordem_de_servico/entities/Pessoa.java
@@ -10,17 +10,17 @@ public abstract class Pessoa {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String nome;
     private String cpf;
     private String telefone;
     private String email;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -59,7 +59,7 @@ public abstract class Pessoa {
     public Pessoa() {
     }
 
-    public Pessoa(Long id, String nome, String cpf, String telefone, String email) {
+    public Pessoa(Integer id, String nome, String cpf, String telefone, String email) {
         this.id = id;
         this.nome = nome;
         this.cpf = cpf;
